Significant changes
The drop command now looks up your held item if you have one and didn't supply an argument.

Added code that adds all Pixelmon to the mod that don't naturally spawn. Before, things such as "Machamp" that don't spawn naturally were not being included in the list of drops. Now, those are all added, with a boolean "isNaturallySpawning" that is checked for spawn-related operations.